slack/requirements.txt
python-dotenv
langchain
openai
slack-sdk 
slack-bolt 
flask
absl-py==1.4.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.0.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
doctran==0.0.9
einops==0.6.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.1
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.3.3
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
inquirerpy==0.3.4
install==1.3.5
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.5
langchain==0.0.279
langcodes==3.3.0
langsmith==0.0.33
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.10
optimum==1.13.1
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandasai==1.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pfzy==0.3.4
phonenumbers==8.13.19
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
preshed==3.0.8
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.0.9
pypdf==3.15.5
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
rsa==4.9
safetensors==0.3.3
scipy==1.11.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.22.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
thinc==8.1.12
tiktoken==0.3.3
tinycss2==1.2.1
tldextract==3.4.4
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.0
transformers-stream-generator==0.0.4
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
widgetsnbextension==4.0.8
xxhash==3.3.0
yarl==1.9.2
